Adds basic concept of an "enhancement tracking receipt"
At some point the KEP process needs to cleanly intersect with the Enhancements
Tracking Process used by the Release Team. My current thinking is that once KEPs
are structured it should be relatively easy to use the Summary of a KEP as a
pointer to its content. These KEP pointers should also be checked into repositories
on GitHub so that we retain the decision history for the project. In order to start
start the transition between a GitHub Issue based workflow to a GitHub Pull Request
based workflow the concept of an "enhancement tracking receipt" which provides the
locations of the information needed by the Release Team to determine whether an
enhancment is able to ride the release train at all:

- a SIG has actually sponsored the work (by accepting a KEP describing the importance and scope of the work)
- a test plan exists (implicity assumed to be a (draft) Pull Request adding or modifying tests)
- documentation exists (implicity assumed to be a (draft) Pull Request adding or modifying documentation)
- a long form release note exists (implicity assumed to be a (draft) Pull Request against the release notes draft)